tunevault/requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.7.1
Django==5.1.3
django-allauth==0.45.0
django-cors-headers==3.10.0
django-redis==5.0.0
djangorestframework
gunicorn==20.1.0
idna==3.10
joblib==1.4.2
kombu==5.4.2
nltk==3.6.5
oauthlib==3.2.2
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
redis==3.5.3
regex==2024.11.6
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
tqdm==4.67.1
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
youtube-dl==2021.12.17
